add item to cart

+++ backend/src/main/java/org/group3/backend/model/Product.java
@@ -24,12 +24,11 @@ public class Product {
     public Product() {
     }
 
-    public double getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
-
 }
